refactor:Separate hotkey related functions
