+++ VimHotkeyWindowManager/HotkeyManager.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <Windows.h>
+#include "definition.h"
+
+void RegisterHotKeys(HWND hWnd);
+
