+++ VimHotkeyWindowManager/main.cpp
@@ -2,6 +2,7 @@
 #include "definition.h"
 #include "MonitorUtils.h"
 #include "WindowManager.h"
+#include "HotkeyManager.h"
 
 
 HMENU hMenu;
@@ -22,14 +23,6 @@ void InitTaskIcon(HWND hWnd) {
 	Shell_NotifyIcon(NIM_ADD, &nid);
 }
 
-void RegisterHotKeys(HWND hWnd) {
-	RegisterHotKey(hWnd, HOTKEY_LEFT, MOD_ALT | MOD_SHIFT, 'H');
-	RegisterHotKey(hWnd, HOTKEY_DOWN, MOD_ALT | MOD_SHIFT, 'J');
-	RegisterHotKey(hWnd, HOTKEY_UP, MOD_ALT | MOD_SHIFT, 'K');
-	RegisterHotKey(hWnd, HOTKEY_RIGHT, MOD_ALT | MOD_SHIFT, 'L');
-	RegisterHotKey(hWnd, HOTKEY_MINIMIZE, MOD_ALT | MOD_SHIFT, 'D');
-	RegisterHotKey(hWnd, HOTKEY_RESTORE, MOD_ALT | MOD_SHIFT, 'U');
-}
 
 bool minimizeWindow(HWND hWnd) {
     minimizedWindow = hWnd;
